avoid use of adjustbox package
Use the normal LaTeX figure environment and not adjustbox,
since adjustbox is not recognised by TtH and results in
incorrect figure numbering/referencing in the HTML output.

The output following this commit is similar but not identical,
since the figures may now float which they didn't before.
Authors should check the output.  There are other possibilities
to sort this out, e.g. don't give some or all of the figures
labelled captions.  The authors can do it differently if they prefer.